requirements.txt
langchain
langchain-openai
langchain-community
langchain-experimental
langgraph
pillow
python-dotenv
tqdm
fastapi
uvicorn[standard]
python-dotenv
gunicorn
langchain-neo4j
neo4j
PyMuPDF
pypdf
langchain-unstructured
markitdown[all]
pdfplumber